+++ backstage/packages/app/src/themes/testTheme.ts
@@ -7,48 +7,48 @@ import {
 } from '@backstage/theme';
 
 const testPalette = {
-      primary: {
-        main: '#343b58',
-      },
-      secondary: {
-        main: '#565a6e',
-      },
-      error: {
-        main: '#8c4351',
-      },
-      warning: {
-        main: '#8f5e15',
-      },
-      info: {
-        main: '#34548a',
-      },
-      success: {
-        main: '#485e30',
-      },
-      background: {
-        default: '#d5d6db',
-        paper: '#d5d6db',
-      },
-      banner: {
-        info: '#34548a',
-        error: '#8c4351',
-        text: '#343b58',
-        link: '#565a6e',
-      },
-      errorBackground: '#8c4351',
-      warningBackground: '#8f5e15',
-      infoBackground: '#343b58',
-      navigation: {
-        background: '#343b58',
-        indicator: '#8f5e15',
-        color: '#d5d6db',
-        selectedColor: '#ffffff',
-      },
-    };
+  primary: {
+    main: '#343b58',
+  },
+  secondary: {
+    main: '#565a6e',
+  },
+  error: {
+    main: '#8c4351',
+  },
+  warning: {
+    main: '#8f5e15',
+  },
+  info: {
+    main: '#34548a',
+  },
+  success: {
+    main: '#485e30',
+  },
+  background: {
+    default: '#d5d6db',
+    paper: '#d5d6db',
+  },
+  banner: {
+    info: '#34548a',
+    error: '#8c4351',
+    text: '#343b58',
+    link: '#565a6e',
+  },
+  errorBackground: '#8c4351',
+  warningBackground: '#8f5e15',
+  infoBackground: '#343b58',
+  navigation: {
+    background: '#343b58',
+    indicator: '#8f5e15',
+    color: '#d5d6db',
+    selectedColor: '#ffffff',
+  },
+};
 
 export const testTheme = createUnifiedTheme({
   ...createBaseThemeOptions({
-    palette: { ...palettes.light, ...testPalette}
+    palette: { ...palettes.light, ...testPalette },
   }),
   defaultPageTheme: 'home',
   fontFamily: 'Comic Sans MS',
